// Users/Gam Fam/Desktop/WEBAPP_03_22_BEFORE_BUG_FIX/src/models/user-dto.ts
export interface UserDTO {
    loginIdentifyer?: string;
    password?: string;
    appVersion?: string;
    deviceId?: string;
    deviceName?: string;
    deviceType?: string;
    mobileNo?: string;
    email?: string;
    transactionId?: string;
    newPassword?: string;
    oldPassword?: string;
}